Fix tracer support for options (was crashing) (#77)
* Fix options not being considered as DU in the tracer

The previous implementation could actually crash:
if the input is Some(DUCase), Fable.Core.Reflection.isUnion returns true
because it does not "see" the option, and only see the wrapped DUCase.
Reflection.FSharpType.GetUnionCases() would then attempt to get the fields
of the input, and since option types are not considered DU under Fable,
it would throw.

* Pin build.fsx dependencies versions to fix the failing build

* Convert array to list only in the tracers unit tests

---------

Co-authored-by: Melvyn Laïly <mlaily@d-edge.com>